+++ src/formatters/functions.rs
@@ -430,7 +430,7 @@ pub fn format_function_args(
                 // parentheses aswell. Otherwise, we just use 1 = opening parentheses.
                 let shape_increment = if hug_table_constructor { 2 } else { 1 };
 
-                let parentheses = format_contained_span(ctx, &parentheses, shape);
+                let parentheses = format_contained_span(ctx, parentheses, shape);
                 let arguments =
                     format_punctuated(ctx, arguments, shape + shape_increment, format_expression);
 
@@ -528,6 +528,61 @@ pub fn format_function_args(
     }
 }
 
+fn should_parameters_format_multiline(
+    ctx: &Context,
+    function_body: &FunctionBody,
+    shape: Shape,
+    block_empty: bool,
+) -> bool {
+    // Check the length of the parameters. We need to format them first onto a single line to check if required
+    let mut line_length = format_singleline_parameters(ctx, function_body, shape)
+        .to_string()
+        .len()
+        + 2; // Account for the parentheses around the parameters
+
+    // If we are in Luau mode, take into account the types
+    // If a type specifier is multiline, the whole parameters should be formatted multiline UNLESS there is only a single parameter.
+    // Otherwise, include them in the total length
+    #[cfg(feature = "luau")]
+    {
+        let (extra_line_length, multiline_specifier_present) = function_body
+            .type_specifiers()
+            .chain(std::iter::once(function_body.return_type())) // Include optional return type
+            .map(|x| {
+                x.map_or((0, false), |specifier| {
+                    let formatted = format_type_specifier(ctx, specifier, shape).to_string();
+
+                    (formatted.len(), formatted.lines().count() > 1)
+                })
+            })
+            .fold(
+                (0, false),
+                |(acc_length, acc_multiline), (length, multiline)| {
+                    (
+                        acc_length + length,
+                        if multiline { true } else { acc_multiline },
+                    )
+                },
+            );
+
+        // One of the type specifiers is multiline, and we have more than one parameter
+        if multiline_specifier_present && function_body.parameters().len() > 1 {
+            return true;
+        }
+
+        // Add the extra length
+        line_length += extra_line_length
+    }
+
+    // If the block is empty, then the `end` will be inlined. We should include this in our line length check
+    if block_empty {
+        line_length += 4 // 4 = " end"
+    }
+
+    let singleline_shape = shape + line_length;
+    singleline_shape.over_budget()
+}
+
 /// Formats a FunctionBody node
 pub fn format_function_body(
     ctx: &Context,
@@ -542,6 +597,13 @@ pub fn format_function_body(
     // If the FunctionBody block is empty, then don't add a newline after the parameters, but add a space:
     // `function() end`
     let block_empty = trivia_util::is_function_empty(function_body);
+
+    #[cfg(feature = "luau")]
+    let generics = function_body
+        .generics()
+        .map(|generic_declaration| format_generic_declaration(ctx, generic_declaration, shape));
+    #[cfg(feature = "luau")]
+    let shape = shape + generics.as_ref().map_or(0, |x| x.to_string().len());
 
     // Check if the parameters should be placed across multiple lines
     let multiline_params = {
@@ -566,42 +628,8 @@ pub fn format_function_body(
             contains_comments || type_specifier_comments
         });
 
-        contains_comments || {
-            // Check the length of the parameters. We need to format them first onto a single line to check if required
-            let types_length: usize;
-            #[cfg(feature = "luau")]
-            {
-                types_length = function_body
-                    .type_specifiers()
-                    .chain(std::iter::once(function_body.return_type())) // Include optional return type
-                    .map(|x| {
-                        x.map_or(0, |specifier| {
-                            format_type_specifier(ctx, specifier, shape)
-                                .to_string()
-                                .len()
-                        })
-                    })
-                    .sum::<usize>()
-            }
-            #[cfg(not(feature = "luau"))]
-            {
-                types_length = 0
-            }
-
-            let mut line_length = format_singleline_parameters(ctx, function_body, shape)
-                    .to_string()
-                    .len()
-                        + 2 // Account for the parentheses around the parameters
-                        + types_length; // Account for type specifiers and return type
-
-            // If the block is empty, then the `end` will be inlined. We should include this in our line length check
-            if block_empty {
-                line_length += 4 // 4 = " end"
-            }
-
-            let singleline_shape = shape + line_length;
-            singleline_shape.over_budget()
-        }
+        contains_comments
+            || should_parameters_format_multiline(ctx, function_body, shape, block_empty)
     };
 
     let (formatted_parameters, mut parameters_parentheses) = match multiline_params {
@@ -649,9 +677,15 @@ pub fn format_function_body(
 
     #[cfg(feature = "luau")]
     {
+        let parameters_shape = if multiline_params {
+            shape.increment_additional_indent()
+        } else {
+            shape
+        };
+
         type_specifiers = function_body
             .type_specifiers()
-            .map(|x| x.map(|specifier| format_type_specifier(ctx, specifier, shape)))
+            .map(|x| x.map(|specifier| format_type_specifier(ctx, specifier, parameters_shape)))
             .collect();
 
         return_type = function_body.return_type().map(|return_type| {
@@ -707,6 +741,7 @@ pub fn format_function_body(
 
     #[cfg(feature = "luau")]
     let function_body = function_body
+        .with_generics(generics)
         .with_type_specifiers(type_specifiers)
         .with_return_type(return_type);
 
@@ -867,24 +902,12 @@ pub fn format_function_declaration(
     .update_leading_trivia(FormatTriviaType::Append(leading_trivia));
     let formatted_function_name = format_function_name(ctx, function_declaration.name(), shape);
 
-    #[cfg(feature = "luau")]
-    let generics = function_declaration
-        .generics()
-        .map(|generic_declaration| format_generic_declaration(ctx, generic_declaration, shape));
-
     let shape = shape + (9 + strip_trivia(&formatted_function_name).to_string().len()); // 9 = "function "
-    #[cfg(feature = "luau")]
-    let shape = shape + generics.as_ref().map_or(0, |x| x.to_string().len());
     let function_body = format_function_body(ctx, function_declaration.body(), true, shape);
 
-    let function_declaration = FunctionDeclaration::new(formatted_function_name)
+    FunctionDeclaration::new(formatted_function_name)
         .with_function_token(function_token)
-        .with_body(function_body);
-
-    #[cfg(feature = "luau")]
-    let function_declaration = function_declaration.with_generics(generics);
-
-    function_declaration
+        .with_body(function_body)
 }
 
 /// Formats a LocalFunction node
@@ -901,25 +924,13 @@ pub fn format_local_function(
     let function_token = fmt_symbol!(ctx, local_function.function_token(), "function ", shape);
     let formatted_name = format_token_reference(ctx, local_function.name(), shape);
 
-    #[cfg(feature = "luau")]
-    let generics = local_function
-        .generics()
-        .map(|generic_declaration| format_generic_declaration(ctx, generic_declaration, shape));
-
     let shape = shape + (6 + 9 + strip_trivia(&formatted_name).to_string().len()); // 6 = "local ", 9 = "function "
-    #[cfg(feature = "luau")]
-    let shape = shape + generics.as_ref().map_or(0, |x| x.to_string().len());
     let function_body = format_function_body(ctx, local_function.body(), true, shape);
 
-    let local_function = LocalFunction::new(formatted_name)
+    LocalFunction::new(formatted_name)
         .with_local_token(local_token)
         .with_function_token(function_token)
-        .with_body(function_body);
-
-    #[cfg(feature = "luau")]
-    let local_function = local_function.with_generics(generics);
-
-    local_function
+        .with_body(function_body)
 }
 
 /// Formats a MethodCall node